% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/femaleProb.R
\name{femaleProb}
\alias{femaleProb}
\title{femaleProb: probability based sex demultiplexing tool}
\usage{
femaleProb(Seuratobj, lognormalized = TRUE, ONLINE = TRUE, xistplots = FALSE, itmax = 100)
}
\arguments{
  \item{Seuratobj}{Seurat object}
  \item{lognormalized}{boolean, whether input data is log-normalized}
  \item{ONLINE}{boolean, whether to fetch Y chromosome genes online from EnsDb}
  \item{xistplots}{boolean, whether to generate diagnostic plots}
  \item{itmax}{integer, iterations for densityMclust fitting before skipping}
}
\description{
Calculates the probability of cells/nuclei belonging to male or female subjects.
Returns a data frame with probabilities based on 4 separate models:
\itemize{
  \item Univariate model, dependent on Xist expression
  \item Multivariate model, dependent on Xist and the sum of Y chromosome genes
  \item Multivariate + nCount, same as above and including unique RNA counts
  \item Xchrom + Ychrom genes, dependent on sum of Xist and Tsix vs the sum of Y chromosome genes
}
}
\author{
Nickolas C. Chu
}
