% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickAdd.R
\name{quickAdd}
\alias{quickAdd}
\title{quickAdd: Quick Sex Assignment for Seurat}
\usage{
quickAdd(Probabilities, Seuratobj)
}
\arguments{
\item{Probabilities}{dataframe of probabilities}

\item{Seuratobj}{A target Seurat object}
}
\description{
Used to add Xist expression, summed Y chromosome gene expression,
probabilities being female, and sex assignments to the metadata of a
Seurat object
}
\author{
Nickolas C. Chu
}
