% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sexAssign.R
\name{sexAssign}
\alias{sexAssign}
\title{sexAssign: Sex assigning function}
\usage{
sexAssign(Probabilities, femalecut = 0.8, malecut = 0.2)
}
\arguments{
\item{Probabilities}{dataframe}

\item{femalecut}{selected probability of being female used to assign females}

\item{malecut}{selected probability of being female used to assign males}
}
\description{
Used to assign sex to each cell based on probability of being female
Default cutoffs:
probability >= 0.8  -> female
probability <= 0.2 -> male
else -> soup
}
\author{
Nickolas C. Chu
}
