% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sxreveal-package.R
\docType{package}
\name{sxreveal-package}
\alias{sxreveal}
\alias{sxreveal-package}
\title{sxreveal: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
Nickolas C. Chu
}
\keyword{internal}
