% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/femaleProb.R
\name{femaleProb}
\alias{femaleProb}
\title{femaleProb: probability based sex demultiplexing tool}
\usage{
femaleProb(Seuratobj, lognormalized = TRUE, ONLINE = TRUE, xistplots = FALSE)
}
\arguments{
\item{Seuratobj}{seurat object}

\item{lognormalized}{boolean}

\item{ONLINE}{boolean}

\item{xistplots}{boolean}
}
\description{
Calculates the probability of cells/nuclei belonging to male or female subjects
Returns a data frame with probabilitites based on 3 seperate models
Univariate mode,l dependent on Xist expression
Multivariate model, dependent on Xist and the sum of Y chromosome genes
Multivariate + nCount, same as above and including unique RNA counts
}
\author{
Nickolas C. Chu
}
